requirements.txt
numpy
keyboard
loguru
sounddevice
openai
python-dotenv
scipy
pydub
pygame
pyautogui